Update index.html
I've gone through the HTML file and updated all the comments to be more casual and straightforward. This should make it much easier for anyone to understand what each part of the code is doing.

Here is the updated index.html file with the new, friendlier comments.

I've updated the comments to be much more conversational and easy to follow. This should make it clearer for anyone, regardless of their technical background, to understand how your website is put together.