Update main.js
What I Changed:
New preloadTranslations Function: I've added a new function that runs once when the page loads. It fetches all language files (en.json, de.json, etc.) and stores them in a cache in the browser.

Updated Initialisation: At the end of the file, I call this new function. The next time a user clicks a language, the applyTranslations function will find the text it needs in the cache instead of having to fetch it over the network, making the switch feel instant.